[front] - feat (AB v2) refactor template tab
