+++ front/components/agent_builder/AgentBuilderRightPanel.tsx
@@ -1,10 +1,7 @@
 import {
-  ArrowPathIcon,
   BarChartIcon,
   Button,
   MagicIcon,
-  Markdown,
-  Page,
   ScrollArea,
   SidebarRightCloseIcon,
   SidebarRightOpenIcon,
@@ -13,14 +10,12 @@ import {
   TabsTrigger,
   TestTubeIcon,
 } from "@dust-tt/sparkle";
-import React, { useContext, useState } from "react";
-import { useFormContext } from "react-hook-form";
+import React, { useState } from "react";
 
-import type { AgentBuilderFormData } from "@app/components/agent_builder/AgentBuilderFormContext";
 import { AgentBuilderPerformance } from "@app/components/agent_builder/AgentBuilderPerformance";
 import { AgentBuilderPreview } from "@app/components/agent_builder/AgentBuilderPreview";
+import { AgentBuilderTemplate } from "@app/components/agent_builder/AgentBuilderTemplate";
 import { usePreviewPanelContext } from "@app/components/agent_builder/PreviewPanelContext";
-import { ConfirmContext } from "@app/components/Confirm";
 import type { FetchAssistantTemplateResponse } from "@app/pages/api/templates/[tId]";
 
 type AgentBuilderRightPanelTabType = "testing" | "performance" | "template";
@@ -143,24 +138,7 @@ function ExpandedContent({
   return (
     <div className="flex min-h-0 flex-1 flex-col">
       {selectedTab === "template" && assistantTemplate && (
-        <div className="flex-1 overflow-y-auto p-4">
-          <div className="flex flex-col gap-4">
-            <TemplateButtons
-              assistantTemplate={assistantTemplate}
-            />
-            {assistantTemplate.helpInstructions && (
-              <>
-                <Markdown content={assistantTemplate.helpInstructions} />
-                {assistantTemplate.helpActions && (
-                  <div className="h-px bg-border" />
-                )}
-              </>
-            )}
-            {assistantTemplate.helpActions && (
-              <Markdown content={assistantTemplate.helpActions} />
-            )}
-          </div>
-        </div>
+        <AgentBuilderTemplate assistantTemplate={assistantTemplate} />
       )}
       {selectedTab === "testing" && (
         <div className="min-h-0 flex-1">
@@ -174,62 +152,6 @@ function ExpandedContent({
           />
         </div>
       )}
-    </div>
-  );
-}
-
-interface TemplateButtonsProps {
-  assistantTemplate: FetchAssistantTemplateResponse;
-}
-
-function TemplateButtons({
-  assistantTemplate,
-}: TemplateButtonsProps) {
-  const confirm = useContext(ConfirmContext);
-  const { setValue } = useFormContext<AgentBuilderFormData>();
-  
-  const handleResetInstructions = async () => {
-    const confirmed = await confirm({
-      title: "Are you sure?",
-      message:
-        "You will lose the changes you have made to the agent's instructions and go back to the template's default settings.",
-      validateVariant: "warning",
-    });
-    
-    if (confirmed && assistantTemplate.presetInstructions) {
-      setValue("instructions", assistantTemplate.presetInstructions);
-    }
-  };
-  
-  const handleResetActions = async () => {
-    const confirmed = await confirm({
-      title: "Are you sure?",
-      message:
-        "You will lose the changes you have made to the agent's tools.",
-      validateVariant: "warning",
-    });
-    
-    if (confirmed) {
-      setValue("actions", []);
-    }
-  };
-  
-  return (
-    <div className="flex items-center justify-end gap-2">
-      <Button
-        label="Reset instructions"
-        onClick={handleResetInstructions}
-        icon={ArrowPathIcon}
-        size="sm"
-        variant="outline"
-      />
-      <Button
-        label="Reset tools"
-        onClick={handleResetActions}
-        icon={ArrowPathIcon}
-        size="sm"
-        variant="outline"
-      />
     </div>
   );
 }
